Initial Commit: 3 programs added.
Created as part of Harvard's CS50 course:
`caesar.c` - A Caesar cipher is applied to a command line argument and returned to the user.
`mario.c` - Prints a pyramid with the size specified by the user who enters this when prompted.
`readability.c` - User prompted for a string, which is analysed and given a readability score which represents the difficulty of reading it.